requirements.txt
Django==1.6.8
Pillow==2.7.0
South==1.0.2
django-extensions==1.4.9
django-model-utils==2.2
ipython==2.3.1
